Rename track module to better convey intent
