+++ packages/arbor-store/src/index.ts
@@ -1,13 +1,13 @@
 export { Arbor, ImmutableArbor } from "./Arbor"
 export { ArrayNodeHandler } from "./ArrayNodeHandler"
+export * from "./decorators"
+export * from "./errors"
+export * from "./guards"
 export { MapNodeHandler } from "./MapNodeHandler"
 export * from "./MutationEngine"
 export { NodeHandler } from "./NodeHandler"
 export { Path } from "./Path"
 export { Subscribers } from "./Subscribers"
-export * from "./decorators"
-export * from "./errors"
-export * from "./guards"
-export * from "./track"
+export * from "./TrackedArbor"
 export type * from "./types"
 export * from "./utilities"
